edtls: verify vouch during handshake
Previously, the server listening on an edtls connection would choose
when to verify a vouch. For example, the CDN server authenticates
uploaders by verifying the vouch on each request. However, on a
long-lived TLS connection the vouch eventually becomes invalid when
the self-signed certificate expires (after 1 hour).

Now a vouch is verified only during the TLS handshake (using the
VerifyPeerCertificate hook in Go 1.8). Servers should now authenticate
requests by calling GetSigningKey on the peer's certificate. This will
work on long-lived TLS connections even after the peer's certificate
expires. To avoid confusion, the Verify function is removed from the
public API.

The cdn package will be updated in a future commit.