cdn: add per-bucket upload permissions
This commit introduces the idea of a CDN bucket, which is an abstraction
over a BoltDB bucket. For example, the CDN bucket "addfriend/1234" is
the keys in the "addfriend" BoltDB bucket prefixed with "1234/". Each
CDN bucket has its own upload permissions specified by the coordinator
when the bucket is created. This will be useful when the mixnet chain is
dynamic and the last server in the chain can change from round to round.

Since upload permissions are now specified at runtime and since we are
trying to move away from the global config file, this commit drops the
-global flag from the alpenhorn-cdn command in favor of specifying the
coordinator key in the CDN's own config file.

This commit also updates the CDN to use the new method of authenticating
edTLS certificates. The coordinator, mixnet, and other packages will be
updated in a future commit.