pkg: use http instead of vrpc for coordination
The commit/reveal protocol for PKG servers now runs over HTTP instead
of vRPC so that PKGs have just one listen address instead of two. This
also adds the CoordinatorClient which should be used by the coordinator
to start new rounds on PKG servers.

The Client and CoordinatorClient use the new edhttp package to send
requests to PKGs. For now, we continue to use one Client object per PKG
even though edhttp supports multiple servers per client.

PKGSettings is renamed to RoundSettings for consistency with the mixnet
package.

The coordinator will be updated in future commits.