mixnet: switch to gRPC for RPCs
The plan is to move toward dynamic configuration for Alpenhorn. With
dynamic configuration the mixnet chain can change from round to round,
which means that mix servers will need more granular authentication than
is provided by vRPC. Instead of extending vRPC, I decided to switch to
gRPC which can authenticate RPCs individually.

The mixnetpb package provides the protobufs used by the mixnet.  With
protobuf and gogoprotobuf installed, the protobuf code was generated
with the command:

    $ protoc -I $GOPATH/src -I. --gofast_out=plugins=grpc:. ./mixnet.proto

Since gRPC does not have an equivalent to net/rpc's RegisterName,
multiple mixing services (AddFriend and Dialing) are now served by a
single Server instance. The new Client type is used by the coordinator
and mix servers to communicate with other mix servers.

Switching to gRPC has not yet had a significant impact on performance.
Here are the results of measuring the end-to-end latency of the dialing
protocol (two mix servers running with mu=1,000,000) with the old
vRPC-based mixnet and the new gRPC-based mixnet:

    name     old time/op  new time/op  delta
    Dialing   45.7s ±11%   46.5s ±10%   ~     (p=0.129 n=24+20)

Note that this commit breaks authentication for the AddOnions and
CloseRound RPCs. This will be fixed in a future commit that adds support
for dynamic mix chains. The Alpenhorn coordinator and client will also
be updated in a future commit.