coordinator: dynamically configure clients
The new AlpenhornConfiguration type is used by the add-friend and
dialing protocols to configure clients. The Alpenhorn administrators
(also called guardians) can update the configuration by posting a new
configuration to /newconfig. The new config must be signed by the
guardians specified in the previous config (and must include the hash of
the previous config). As a result, a sequence of config updates forms a
hash chain. This enables clients to catch-up (after being offline for
awhile) to a new config from an earlier config in the chain.

At the start of every round, the coordinator broadcasts the hash of the
configuration for the round to all clients. If this hash differs from
the hash of the config used in previous rounds, a client can download
the new config by sending a request to /config, to which the coordinator
responds with a chain of configs. The client should verify the chain
starting from the previous config before continuing the round. This
verification step can include alerting the user about changes that
potentially reduce security (e.g., removing a PKG, mix server, or
guardian).

Note that this scheme does not guarantee freshness of configs. The
coordinator can withhold config updates in an attempt to get clients to
use compromised servers. There are multiple ways to fix this problem:
short expiration times on configs, gossiping, or regular endorsements of
the current config. We have not yet decided on a plan for freshness.

With dynamic configuration, the "global config" is no longer needed, so
the -global flag is removed from the coordinator command. Instead, the
coordinator should be initialized with the -bootstrap flag, which
specifies the starting configs in the add-friend and dialing
configuration chains.

This commit also makes the coordinator compatible with earlier changes
to the mixnet and CDN. Other notable changes include:

    * The coordinator now listens on edtls, which means the typesocket
    package needs to be updated.

    * The websocket paths for add-friend and dialing have changed:

        /afws -> /addfriend/ws
        /dws -> /dialing/ws

    * Persistence is improved.

The client will be updated in a future commit.