client: adapt to dynamic configuration
This updates the client to bring it in sync with the last several
commits. In particular, the client is now dynamically configured, so the
only connection settings needed upfront are the coordinator's address
and key (for edtls). Information about the mixchain and PKG servers is
obtained from the coordinator in the "newround" message handler, which
downloads and verifies the add-friend/dialing config for that round.

The client needs starting add-friend and dialing configs to bootstrap
trust in the system. The Bootstrap function sets the starting configs
and should only be called once when the client is created. The Bootstrap
function is the only way to set the client's add-friend and dialing
configs without verifying signatures; the starting configs are assumed
to have been verified out-of-band. In practice, the starting configs
would be packaged with the software.

The client now uses the same login key for all PKGs, instead of a unique
key per PKG. This is so that new PKGs can pre-populate their users
database to avoid an attack described in an earlier commit. When a new
PKG server is added to the add-friend protocol, the client tries to
login to the PKG before attempting to register.

Since PKGs can change from round to round, the IncomingFriendRequest
type now has a Verifiers field that lists the PKGs used to verify the
friend request. This is useful in tests and so that users can make an
informed decision when approving long-lived friend requests.

This commit also updates the end-to-end test in alpenhorn_test.go and
tests adding a new PKG to the add-friend protocol.