pkg: add status handler
Now clients can check their account status at any time, without having
to extract a private key. This is especially useful when a new PKG
server comes online. Suppose that users are required to register with a
new PKG when it comes online. This gives an attacker the opportunity to
lock a user out of Alpenhorn every time a new PKG is added to the system
by compromising the user's email address. Instead, a new PKG should
pre-populate its users database using the databases of existing PKGs,
and the user should confirm their account exists on the new PKG using
the client's CheckStatus() function.