mixnet: make mix chains dynamic
The mix chain is now dynamic and provided by the coordinator in the
NewRound RPC. The mix chain is stored in the round state so that each
round can have a different chain. The mixers make no effort to verify
that the coordinator provided the right chain; this verification is left
to the clients. In the future, we may have mixers refuse to participate
in old chains to avoid issues with stale configs.

This commit fixes authentication in the AddOnions and CloseRound RPCs.
This also drops the -global flag from the alpenhorn-mixer command in
favor of specifying the coordinator key in the mix server's config file.