log: new package for structured logging
The log package is inspired by logrus but aims to fix some of the
awkwardness in the logrus API. In particular, logger.WithFields returns
a Logger instead of an Entry. The zap package (go.uber.org/zap) has a
similar API to ours but is harder to customize. Speed is not the primary
goal of this package but the performance seems on par with logrus.