config: repurpose package for general config server
The Vuvuzela coordinator will need a way to manage configuration
updates. Instead of duplicating a bunch of code from the Alpenhorn
coordinator, we generalized dynamic configuration in the config package.

The new SignedConfig type generalizes the AlpenhornConfig type. Its
Inner field points to a protocol-specific configuration (e.g., for the
add-friend, dialing, or convo protocol). Inner config types must be
registered with the config package so that signed configs can be
unmarshaled from JSON seamlessly.

The new Server type generalizes the config functionality currently
provided by the coordinator. It handles queries for the current config
from clients and updates to the config from guardians. The coordinator
will be updated to use config.Server in a future commit.

The config server has its own persistence plan separate from the
coordinator. This seems simpler than trying to store the configs in the
coordinator's state.