coordinator: use the new config server
All of the dynamic configuration functionality has moved to the config
package. The coordinator now has an underlying config server that
handles dynamic configuration. The config server is loaded from the
ConfigServerPersistPath, which must be specified when running the
coordinator. The path must already exist and have been created using
config.CreateServerState (which replaces the Bootstrap function).

This commit also revamps the init process for the alpenhorn-coordinator
command to accommodate the new persist paths. The format of the bootstrap
config is still in flux. We will likely switch from TOML to JSON since
our toml package can't decode an arbitrary SignedConfig (we would need
something like json.RawMessage).

The client will be updated in an upcoming commit.