all: separate config server from coordinator
Previously, the coordinator was the config server, so clients had to
learn the coordinator's key and address out-of-band. That also made it
difficult to update the coordinator's key and address.

Now there is a single config server (independent of the coordinator)
that provides the configurations for all the services ("AddFriend",
"Dialing", and "Convo"). The URL of this config server is hardcoded so
that clients can get the latest configs without any setup.

The coordinator's key and address are now part of the AddFriend/Dialing
configs and can be updated like any other server. The Alpenhorn client
no longer takes the coordinator info as a parameter. Instead, the client
learns the current coordinator by fetching the latest config when
connecting. This will also simplify the bootstrap process for clients.

The config server is not trusted, assuming users verify the guardian
keys in the initial config they receive from the server. The config
server also runs on LetsEncrypt-backed HTTPS instead of edTLS so that
clients don't need to obtain the config server's key out-of-band.

The coordinator bootstrap process is replaced with the -setConfig flag
in the alpenhorn-config-server command.

Also, some routes and APIs have changed along the way.