config: require configs to be signed by listed guardians
Previously, VerifyConfigChain checked only that the current config was
signed by the previous config's guardians. Now it additionally checks
that the current config is also signed by all the guardians listed in
the config.

This also adds the Verify method, which checks that a config is signed
by all the guardians listed in the config. Verify is a weaker check than
VerifyConfigChain, so users of the config package should use
VerifyConfigChain when a previous config is available.

When fetching the current config, the config client now validates the
config, checks expiration dates, and verifies the config with Verify.