client: enable user to update a call's intent before sending it
In our plan for forward secrecy for the convo protocol, the Vuvuzela
client must be able to update a call's intent right before it is sent.

This replaces the SentCall handler (which was called after a call was
sent) with the SendingCall handler (which is called right before a call
is sent). Inside this handler, the application can call UpdateIntent to
change a call's intent and then call SessionKey to get the call's
session key. The call's intent can not be changed after obtaining its
session key.

This also removes the named intent constants because the application can
choose its own semantics for intents.

Updates vuvuzela/vuvuzela#3.