typesocket: expose client read loop via Serve method
Previously there wasn't a good way to tell when a client's connection
had failed because the read loop was hidden behind a Go routine. Now the
Serve method runs the read loop and always returns a non-nil error when
the loop exits (similar to http.Serve). This gives applications a way to
learn about read errors.

This is the first commit in a series of commits that enable the Vuvuzela
client to reconnect when the connection is lost.