client: expose disconnects to application
This makes several improvements to how the client connects to the
coordinator so that applications can detect disconnects and then
reconnect.

For simplicity, the Connect method is now split into two methods,
ConnectAddFriend and ConnectDialing. Some initialization code has also
moved out of the Connect methods. Notably, the keywheel is now loaded in
LoadClient instead of at connect time.

The Connect methods return a channel that receives an error message (the
error returned from typesocket.(*ClientConn).Serve) when the connection
fails. Applications can block reading on this channel before attempting
to reconnect. An alternative design is to add another method to the
client (e.g., ServeAddFriend) that calls the typesocket Serve method.
Returning a channel from Connect seems cleaner.