all: write server logs to disk
The server commands now write logs to disk (in JSON-line format) so that
logs can be processed by tools like filebeat. Errors are written both to
disk and to stderr. If the logsDir config parameter isn't set, then
servers will write all log entries to stderr (in text format).

Since servers will write logs to stderr less often, the text format now
includes a date for each log entry. This removes the need for the
dateLogger.

The OutputDir entry handler in the log package implements logging to a
directory. To reduce allocations, the (*log.Entry).JSON method now takes
a writer as an argument.