pkg: track each user's last successful extraction
To make account recovery possible, the PKGs will have a lockout policy
that allows an account to be recovered only if it has not logged in
(i.e., successfully extracted a private key) for 30 days. To implement
lockouts, the PKG must track each user's last successful extraction.

For every extraction, the PKG must write the timestamp of the extraction
to disk before returning to the user. Otherwise, consider a user that
logs in only once every 30 days to keep their account active. If the PKG
returns a key to the user before writing to disk and then crashes, an
attacker can steal the user's account: the user thinks everything is
fine and wont login again for 30 days, but really their account has
expired and can be recovered by an attacker.

We use Badger to record extraction timestamps because it has much higher
write throughput than we are able to get with Postgres. With my laptop's
SSD, Badger can do 1M writes in 6 seconds. Also, with Badger we don't
need to implement our own batching since Badger already batches
concurrent db.Update calls.