pkg: switch to using email addresses as usernames
Our plan for account recovery (after a user loses their long-term
private key) was to allow a user to reset their account if they haven't
extracted a key in 30 days. However, with first-come-first-serve
usernames, there is no way to alert the user or verify their reset
request. This could lead to usernames being stolen if a user goes
offline for an extended period.

With email addresses, there is a clearer recovery plan: we can verify
that the user owns their email address (by sending a confirmation email)
before resetting the account.

The new flow for creating an account is as follows:

1. The user signs up with the registrar (https://vuvuzela.io).
2. The registrar sends a preregistration request to each PKG. This
causes each PKG to send a verification email to the user.
3. Each email contains a token that the user must paste into the
Vuvuzela client to complete the account registration. This step
associates the long-term key of the Vuvuzela client with the user's
email address.

In the future we may use DKIM or a local webserver to avoid having to
copy and paste from multiple emails.

The PKGs send email through an SMTP relay that should be configured in
the PKG's config file.

This commit also tweaks the PKG's HTTP routes.

See issue vuvuzela/vuvuzela#35.